+++ example/angular2-try/components/app.ts
@@ -22,7 +22,7 @@ import {CommonLogin} from '../directives/common.login';
     directives: [RouterOutlet, RouterLink]
 })
 @RouteConfig([
-    {path: '/', as:'index', component:Start},
+    {path: '/', as:'index', component:Ssq},
     {path: '/login', as:'login', component:Login},
     {path: '/start', as:'start', component:Start},
     {path: '/zhuanpan', as: 'zhuanpan', component:Zhuanpan},