+++ example/angular2-try/components/ssq.js
@@ -128,6 +128,10 @@ var __param = (this && this.__param) || function (paramIndex, decorator) {
                 Ssq.prototype.betConfirm = function () {
                     var betPlanContent = this.commonService.betPlanContent;
                     betPlanContent.lottery = 'ssq';
+                    //request
+                    utils_1.utils.fetch("http://localhost:8888/json/message2", function (jsonData) {
+                        console.log(jsonData.result.userName);
+                    }, { 'id': 999, 'name': '12345' }, 'POST');
                     if (this.betInfo.get('betCount') < 1) {
                         this.router.navigate('/ssqBetConfirm');
                         return;