+++ History.md
@@ -1,6 +1,7 @@
 3.0.5-rc1/ 2015-08-13
 ==================
 * add: add 'common-login' angular2 directives
+* add: add window.fetch polyfill in perfmjs/utils
 
 3.0.4/ 2015-08-12
 ==================